+++ assets/js/angular/directives/mediapreview.js
@@ -22,7 +22,9 @@
                                     url = url.replace("site:", COCKPIT_SITE_BASE_URL)
                                 }
 
-                                $r = '<div class="media-url-preview" style="background-image:url('+encodeURI(url)+')"></div>';
+                                var style = elm.attr('style') || '';
+
+                                $r = '<div class="media-url-preview" style="background-image:url('+encodeURI(url)+');background-size:cover;'+style+'"></div>';
                             }
 
                             if (url.match(/\.(mp4|mpeg|ogv|webm|wmv)$/i)) {